Rename module
Renamed the `coordinate` module to `coordinates`, since it is more readable using the `from bidimensional.coordinates` import statement structure.
