+++ src/bidimensional/polygons.py
@@ -2,7 +2,7 @@ from matplotlib import pyplot as plt
 
 import bidimensional.operations as op
 from bidimensional.circumcircle import Circumcircle
-from bidimensional.coordinate import Coordinate2D
+from bidimensional.coordinates import Coordinate2D
 
 
 class Triangle:
